Move the unification code to a separate file
