+++ compiler/noirc_frontend/src/hir_def/types/arithmetic.rs
@@ -3,7 +3,7 @@ use std::collections::BTreeMap;
 use acvm::{AcirField, FieldElement};
 use noirc_errors::Location;
 
-use crate::{BinaryTypeOperator, Type, TypeBindings, UnificationError};
+use crate::{BinaryTypeOperator, Type};
 
 impl Type {
     /// Try to canonicalize the representation of this type.
@@ -327,145 +327,6 @@ impl Type {
             }
             _ => None,
         }
-    }
-
-    /// Try to unify equations like `(..) + 3 = (..) + 1`
-    /// by transforming them to `(..) + 2 =  (..)`
-    pub(super) fn try_unify_by_moving_constant_terms(
-        &self,
-        other: &Type,
-        bindings: &mut TypeBindings,
-    ) -> Result<(), UnificationError> {
-        if let Type::InfixExpr(lhs_a, op_a, rhs_a, _) = self {
-            if let Some(inverse) = op_a.approx_inverse() {
-                let kind = lhs_a.infix_kind(rhs_a);
-                let dummy_location = Location::dummy();
-                if let Ok(rhs_a_value) = rhs_a.evaluate_to_field_element(&kind, dummy_location) {
-                    let rhs_a = Box::new(Type::Constant(rhs_a_value, kind));
-                    let new_other =
-                        Type::inverted_infix_expr(Box::new(other.clone()), inverse, rhs_a);
-
-                    let mut tmp_bindings = bindings.clone();
-                    if lhs_a.try_unify(&new_other, &mut tmp_bindings).is_ok() {
-                        *bindings = tmp_bindings;
-                        return Ok(());
-                    }
-                }
-            }
-        }
-
-        if let Type::InfixExpr(lhs_b, op_b, rhs_b, inversion) = other {
-            if let Some(inverse) = op_b.approx_inverse() {
-                let kind = lhs_b.infix_kind(rhs_b);
-                let dummy_location = Location::dummy();
-                if let Ok(rhs_b_value) = rhs_b.evaluate_to_field_element(&kind, dummy_location) {
-                    let rhs_b = Box::new(Type::Constant(rhs_b_value, kind));
-                    let new_self =
-                        Type::InfixExpr(Box::new(self.clone()), inverse, rhs_b, !inversion);
-
-                    let mut tmp_bindings = bindings.clone();
-                    if new_self.try_unify(lhs_b, &mut tmp_bindings).is_ok() {
-                        *bindings = tmp_bindings;
-                        return Ok(());
-                    }
-                }
-            }
-        }
-
-        Err(UnificationError)
-    }
-
-    /// Try to unify the following equations:
-    /// - `A + rhs = other` -> `A = other - rhs`
-    /// - `A - rhs = other` -> `A = other + rhs`
-    /// - `lhs + B = other` -> `B = other - lhs`
-    /// - `lhs - B = other` -> `B = lhs - other`
-    /// - `other = A + rhs` -> `A = other - rhs`
-    /// - `other = A - rhs` -> `A = other + rhs`
-    /// - `other = lhs + B` -> `B = other - lhs`
-    /// - `other = lhs - B` -> `B = lhs - other`
-    pub(super) fn try_unify_by_isolating_an_unbound_type_variable(
-        &self,
-        other: &Type,
-        bindings: &mut TypeBindings,
-    ) -> Result<(), UnificationError> {
-        self.try_unify_by_isolating_an_unbound_type_variable_in_self(other, bindings).or_else(
-            |_| other.try_unify_by_isolating_an_unbound_type_variable_in_self(self, bindings),
-        )
-    }
-
-    /// Try to unify the following equations:
-    /// - `A + rhs = other` -> `A = other - rhs`
-    /// - `A - rhs = other` -> `A = other + rhs`
-    /// - `lhs + B = other` -> `B = other - lhs`
-    /// - `lhs - B = other` -> `B = lhs - other`
-    fn try_unify_by_isolating_an_unbound_type_variable_in_self(
-        &self,
-        other: &Type,
-        bindings: &mut TypeBindings,
-    ) -> Result<(), UnificationError> {
-        if let Type::InfixExpr(lhs_lhs, lhs_op, lhs_rhs, _) = self {
-            let lhs_op_inverse = lhs_op.inverse();
-
-            // Check if it's `A + rhs = other` or `A - rhs = other`
-            if let (Some(op_a_inverse), Type::TypeVariable(lhs_lhs_var)) =
-                (lhs_op_inverse, lhs_lhs.as_ref())
-            {
-                if lhs_lhs_var.1.borrow().is_unbound() {
-                    // We can say that `A = other - rhs` or `A = other + rhs` respectively
-                    let new_rhs =
-                        Type::infix_expr(Box::new(other.clone()), op_a_inverse, lhs_rhs.clone());
-
-                    let mut tmp_bindings = bindings.clone();
-                    if lhs_lhs.try_unify(&new_rhs, &mut tmp_bindings).is_ok() {
-                        *bindings = tmp_bindings;
-                        return Ok(());
-                    }
-                }
-            }
-
-            // Check if it's `lhs + B = other`
-            if let (BinaryTypeOperator::Addition, Type::TypeVariable(lhs_rhs_var)) =
-                (lhs_op, lhs_rhs.as_ref())
-            {
-                if lhs_rhs_var.1.borrow().is_unbound() {
-                    // We can say that `B = other - lhs`
-                    let new_rhs = Type::infix_expr(
-                        Box::new(other.clone()),
-                        BinaryTypeOperator::Subtraction,
-                        lhs_lhs.clone(),
-                    );
-
-                    let mut tmp_bindings = bindings.clone();
-                    if lhs_rhs.try_unify(&new_rhs, &mut tmp_bindings).is_ok() {
-                        *bindings = tmp_bindings;
-                        return Ok(());
-                    }
-                }
-            }
-
-            // Check if it's `lhs - B = other`
-            if let (BinaryTypeOperator::Subtraction, Type::TypeVariable(lhs_rhs_var)) =
-                (lhs_op, lhs_rhs.as_ref())
-            {
-                if lhs_rhs_var.1.borrow().is_unbound() {
-                    // We can say that `B = lhs - other`
-                    let new_rhs = Type::infix_expr(
-                        lhs_lhs.clone(),
-                        BinaryTypeOperator::Subtraction,
-                        Box::new(other.clone()),
-                    );
-
-                    let mut tmp_bindings = bindings.clone();
-                    if lhs_rhs.try_unify(&new_rhs, &mut tmp_bindings).is_ok() {
-                        *bindings = tmp_bindings;
-                        return Ok(());
-                    }
-                }
-            }
-        }
-
-        Err(UnificationError)
     }
 }
